+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Removed
 
+- Drop returning the page UID in `createFakeFrontend` (#2383)
 - Drop `ExtbaseConfiguration` (#2379)
 - Drop the `Time` interface (#2378)
 - Drop `FrontEndUserMapper::findByUserName()` (#2377)